refactor(sql): rewrite the package relationship
This commit split the `pkg_relation` table into two table:
`outdated_dep` and `missing_dep`. This approach is better because this
is a one-to-many relationship. The old implementation turn that
relationship to many-to-many which is error-prune during development.

This commit also add `id` for `mark` table to easily trace back details
of a mark. A new module `model` is also created for `ORM` like database
operation in future design.

Signed-off-by: Avimitin <avimitin@gmail.com>